A fix for Bug#850031 "Assertion failure in iproto_interact()"
A fix for Bug#850031 "Assertion failure in iproto_interact()".

Fix fiber_bread() to read at_least bytes, not
buf->len - at_least bytes, as it used to work before.
Many places that call fiber_bread rely on this.

+++ include/fiber.h
@@ -148,7 +148,7 @@ void yield(void);
 void fiber_destroy_all();
 
 struct msg *read_inbox(void);
-int fiber_bread(struct tbuf *, size_t v);
+ssize_t fiber_bread(struct tbuf *, size_t v);
 
 inline static void iov_add_unsafe(const void *buf, size_t len)
 {
